-cleanup for file rename script to accept an arg instead of hardcoding the new replacement value -video manifest download script now setup to download all video manifests in the playlist with a single use of this script, as it now auto clicks, loads, and iterates through each of the videos in the playlist

+++ shell/renameAccountingFilesInDirectory.sh
@@ -1,11 +1,22 @@
 #!/usr/local/bin/bash
 
 DIR=$1
+REPLACEMENT=$2
+if [[ "$DIR" == "" ]]; then
+  echo "missing directory arg1"
+  exit 1
+fi
+if [[ "$REPLACEMENT" == "" ]]; then
+  echo "missing text replacement arg2"
+  exit 1
+fi
+
+cd "$DIR"
 FILES=($( ls "$DIR" ))
 for ((i=0; i<${#FILES[@]}; i++)); do
   FILE_NAME="${FILES[$i]}"
   REG_MATCH="[a-zA-Z0-9]{5,9}"
-  REPLACEMENT="acctg6000"
   NEW_NAME=$(echo "$FILE_NAME" | sed -E "s|$REG_MATCH|$REPLACEMENT|")
+  echo "$NEW_NAME"
   mv "${FILES[$i]}" "$NEW_NAME"
 done
